+++ test/ex_check/project_cases/deps_test.exs
@@ -1,4 +1,4 @@
-defmodule ExCheck.ProjectCases.RunAfterTest do
+defmodule ExCheck.ProjectCases.DepsTest do
   use ExCheck.ProjectCase, async: true
 
   @a_eval """
@@ -18,16 +18,16 @@ defmodule ExCheck.ProjectCases.RunAfterTest do
   @config """
   [
     tools: [
-      {:c, command: ["elixir", "-e", #{inspect(@c_eval)}], run_after: [:a, :b]},
+      {:c, command: ["elixir", "-e", #{inspect(@c_eval)}], deps: [:a, :b]},
       {:a, command: ["elixir", "-e", #{inspect(@a_eval)}], order: 1},
       {:b, command: ["elixir", "-e", #{inspect(@b_eval)}]},
-      {:d, command: "dont_matter", run_after: [:e]},
-      {:e, command: "dont_matter_too", run_after: [:d]}
+      {:d, command: "dont_matter", deps: [:e]},
+      {:e, command: "dont_matter_too", deps: [:d]}
     ]
   ]
   """
 
-  test "run_after", %{project_dir: project_dir} do
+  test "deps", %{project_dir: project_dir} do
     config_path = Path.join(project_dir, ".check.exs")
     File.write!(config_path, @config)
 
@@ -36,14 +36,14 @@ defmodule ExCheck.ProjectCases.RunAfterTest do
     assert String.contains?(output, "a success")
     assert String.contains?(output, "b success")
     assert String.contains?(output, "c success")
-    assert String.contains?(output, "d skipped due to broken tool dependency e")
-    assert String.contains?(output, "e skipped due to broken tool dependency d")
+    assert String.contains?(output, "d skipped due to unsatisfied dependency e")
+    assert String.contains?(output, "e skipped due to unsatisfied dependency d")
     assert String.contains?(output, "a_out-b_out")
 
     assert String.match?(output, ~r/running b.*running a.*running c/s)
   end
 
-  test "run_after (no parallel)", %{project_dir: project_dir} do
+  test "deps (no parallel)", %{project_dir: project_dir} do
     config_path = Path.join(project_dir, ".check.exs")
     File.write!(config_path, @config)
 
@@ -52,8 +52,8 @@ defmodule ExCheck.ProjectCases.RunAfterTest do
     assert String.contains?(output, "a success")
     assert String.contains?(output, "b success")
     assert String.contains?(output, "c success")
-    assert String.contains?(output, "d skipped due to broken tool dependency e")
-    assert String.contains?(output, "e skipped due to broken tool dependency d")
+    assert String.contains?(output, "d skipped due to unsatisfied dependency e")
+    assert String.contains?(output, "e skipped due to unsatisfied dependency d")
     assert String.contains?(output, "a_out-b_out")
 
     assert String.match?(output, ~r/running b.*running a.*running c/s)